requirements.txt
python-dotenv
requests
langchain-openai
browser-use
playwright
keyring 